CMN-293: opens details dialog from group list
