+++ src/apps/loan-list/modal/renew-loans-modal-content.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, FC, useCallback } from "react";
 import { useIntersection } from "react-use";
 import { useQueryClient } from "react-query";
-import SelectableMaterial from "../materials/selectable-material";
+import SelectableMaterial from "../materials/selectable-material/selectable-material";
 import { useRenewLoansV2, getGetLoansV2QueryKey } from "../../../core/fbs/fbs";
 import {
   getAmountOfRenewableLoans,
